+++ src/lib/dialog.lua
@@ -12,7 +12,6 @@ function Dialog:__call(player)
     local instance = {
         object_ = cj.DialogCreate(),
         owner_ = player,
-        buttons_ = {},
     }
 
     setmetatable(instance, self)
@@ -27,23 +26,29 @@ function mt:Remove()
 
     self.object_ = nil
     self.owner_ = nil
-    self.buttons = nil
     self = nil 
 end
 
 function mt:Clear()
     cj.DialogClear(self.object_)
-    self.buttons_ = {}
+    
+    local pairs = pairs
+    for key, btn in pairs(self) do 
+        if (key ~= 'object_') and (key ~= 'owner_') then
+            btn = nil
+            self[key] = nil
+        end
+    end
 end
 
 -- label是索引，不填會預設為按鈕文字
 function mt:AddButton(text, label, hotkey)
     local key = label or text
-    self.buttons_[key] = cj.DialogAddButton(self.object_, text, hotkey or 0)
+    self[key] = cj.DialogAddButton(self.object_, text, hotkey or 0)
 end
 
 function mt:FindButton(key)
-    return self.buttons_[key]
+    return self[key]
 end
 
 function mt:Show(is_show)